requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==1.10.15
pandas==2.2.2
scikit-learn==1.5.2
joblib==1.4.2
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
slowapi==0.1.9
limits==3.12.0
redis==5.0.8